--- migration/tasklist-migration/src/main/java/io/camunda/migration/tasklist/TasklistMigrationProperties.java
@@ -1,34 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.migration.tasklist;
-
-import io.camunda.search.connect.configuration.ConnectConfiguration;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "camunda.migration.tasklist")
-public class TasklistMigrationProperties {
-
-  private int batchSize = 20;
-  private ConnectConfiguration connect = new ConnectConfiguration();
-
-  public ConnectConfiguration getConnect() {
-    return connect;
-  }
-
-  public void setConnect(final ConnectConfiguration connect) {
-    this.connect = connect;
-  }
-
-  public int getBatchSize() {
-    return batchSize;
-  }
-
-  public void setBatchSize(final int batchSize) {
-    this.batchSize = batchSize;
-  }
-}
